requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==11.1.0
PyJWT==2.10.1
dj-database-url==2.1.0
requests==2.32.3